+++ client/package.json
@@ -17,6 +17,7 @@
     "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
     "chart.js": "^4.4.2",
+    "crypto-js": "^4.2.0",
     "emailjs-com": "^3.2.0",
     "mailgun-js": "^0.22.0",
     "react": "^18.2.0",
