+++ server/app/Http/Controllers/EmployeeAuthController.php
@@ -147,14 +147,14 @@ class EmployeeAuthController extends Controller
         ]);
     }
 
-    public function checkEmail($email)
-    {
-        $employee = Employee::where('email', $email)->first();
+    // public function checkEmail($email)
+    // {
+    //     $employee = Employee::where('email', $email)->first();
 
-        if ($employee) {
-            return response()->json(['success' => true, 'employee' => $employee]);
-        } else {
-            return response()->json(['success' => false, 'message' => 'Email not found'], 404);
-        }
-    }
+    //     if ($employee) {
+    //         return response()->json(['success' => true, 'employee' => $employee]);
+    //     } else {
+    //         return response()->json(['success' => false, 'message' => 'Email not found'], 404);
+    //     }
+    // }
 }
